main.py
Class bullet : The Bullet class inherits from turtle.Turtle. It is used to represent projectiles that the player can fire. By default, these bullets are invisible and are only activated when fired.

Fire : This method allows you to shoot a ball. If the ball is inactive, it is placed above the player, made visible, and its status is marked as "active".

Move : This method handles the vertical movement of the ball. It also checks if the ball goes off-screen, in which case it is disabled and made invisible.

Class Enemy : The Enemy class represents each alien. It uses a specific image (alien.gif) and is positioned according to coordinates given when it is created.

Creating the Enemy Grid : A grid of 5 rows and 10 columns is created by calculating the initial position of each enemy based on its index in the grid. Enemies are spaced evenly.

Enemy Movement Variables : These variables control the speed and direction of enemy movement. enemy_dx adjusts horizontal movement, while enemy_dy is used to move enemies downward when they change direction. enemy_delay adjusts the overall speed of enemies.